Fixed rendering bugs, but now on to fixing physics bugs

+++ ProjectBane/Bane-Graphics-Core/Graphics/Data/CameraData.h
@@ -7,7 +7,7 @@ typedef struct CAMERA_DATA {
 	IRenderPassInfo* CurrentRenderPass;
 	float4x4 View;
 	float4x4 Projection;
-	float3 Position;
+	double3 Position;
 	float ZResolution;
 	float FarPlane;
 } CAMERA_DATA;
